Changes:
Defined the function x so that the path is chosen randomly.
Change in the definition of a bifurcation. Now the end sensors are required to detect above the threshold (i.e. detect black) and the middle one is required to detect below the threshold (white).
Other minor changes.

Changes committed:
	modified:   line_following.py